+++ app/models/topic.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-# Topics group together discussions on the forum.
+# Allows users to start new discussion topics on the forum.
 class Topic < ActiveRecord::Base
   attr_accessor :description
 
@@ -13,29 +13,27 @@ class Topic < ActiveRecord::Base
   # Named Scopes
   scope :not_banned, -> { joins(:user).merge(User.where(banned: false)) }
   def self.search(arg)
-    where(
-      'topics.name ~* ? or topics.id in (select comments.topic_id from comments where comments.description ~* ? and comments.deleted = ?)',
-      arg.to_s.split(/\s/).collect { |l| l.to_s.gsub(/[^\w\d%]/, '') }.collect { |l| "(\\m#{l})" }.join('|'),
-      arg.to_s.split(/\s/).collect { |l| l.to_s.gsub(/[^\w\d%]/, '') }.collect { |l| "(\\m#{l})" }.join('|'),
-      false
-    )
+    # TODO: Implement full text search through PgSearch multisearchable
+    where('1 = 1')
   end
 
   # Model Validation
-  validates :name, :user_id, presence: true
-  validates :name, length: { maximum: 40 }
+  validates :title, :slug, :user_id, presence: true
+  validates :title, length: { maximum: 40 }
   validates :description, presence: true, if: :new_record?
+  validates :slug, uniqueness: { scope: :deleted }
+  validates :slug, format: { with: /\A(?!\Anew\Z)[a-z][a-z0-9\-]*\Z/ }
 
   # Model Relationships
-  has_many :comments
   belongs_to :user
+  has_many :comments, -> { order :id }
   has_many :subscriptions
   has_many :subscribers, -> { current.where(emails_enabled: true).where(subscriptions: { subscribed: true }) }, through: :subscriptions, source: :user
   has_many :topic_tags
   has_many :tags, -> { where(deleted: false).order(:name) }, through: :topic_tags
 
   def to_param
-    "#{id}-#{name.parameterize}"
+    slug_was.to_s
   end
 
   def editable_by?(current_user)
@@ -59,7 +57,7 @@ class Topic < ActiveRecord::Base
   private
 
   def create_first_comment
-    comments.create(description: description, user_id: user_id)
+    comments.create description: description, user_id: user_id if description.present?
     get_or_create_subscription(user)
   end
 end