+++ db/migrate/20160613200918_rename_name_to_title_for_topics.rb
@@ -1,0 +1,5 @@
+class RenameNameToTitleForTopics < ActiveRecord::Migration
+  def change
+    rename_column :topics, :name, :title
+  end
+end