+++ test/models/topic_test.rb
@@ -7,7 +7,8 @@ class TopicTest < ActiveSupport::TestCase
   def topic_params
     {
       user_id: User.first.id,
-      name: 'Topic Name',
+      title: 'Topic Title',
+      slug: 'topic-title',
       description: 'First Comment'
     }
   end
@@ -20,21 +21,21 @@ class TopicTest < ActiveSupport::TestCase
       end
     end
     assert_equal 0, topic.errors.count
-    assert_equal 'Topic Name', topic.name
+    assert_equal 'Topic Title', topic.title
     assert_equal User.first.id, topic.user.id
     assert_equal 'First Comment', topic.comments.first.description
     assert_equal User.first.id, topic.comments.first.user_id
   end
 
-  test 'should not save topic with blank name' do
-    topic = Topic.new(topic_params.merge(name: ''))
+  test 'should not save topic with blank title' do
+    topic = Topic.new(topic_params.merge(title: ''))
     assert_difference('Comment.count', 0) do
       assert_difference('Topic.count', 0) do
         topic.save
       end
     end
     assert_equal 1, topic.errors.count
-    assert_equal ["can't be blank"], topic.errors[:name]
+    assert_equal ["can't be blank"], topic.errors[:title]
   end
 
   test 'should not save topic without user' do